most regresion tests now merged into suite
