--- beast/tools/tests/test_beast_settings.py
@@ -1,258 +1,0 @@
-from tempfile import NamedTemporaryFile
-from beast.tools import beast_settings  # , get_libfiles
-from astropy.tests.helper import remote_data
-
-
-@remote_data
-def test_beast_settings():
-    """
-    Test that a given text file creates the expected beast_examples class.
-
-    Text is copied over from the phat_small example in beast-examples.
-    """
-
-    # make a temp file to hold the settings text file
-    temp_file = NamedTemporaryFile(suffix=".txt")
-    # also need to download some beast library files
-    # get_libfiles.get_libfiles()
-
-    with open(temp_file.name, "w") as beast_file:
-
-        beast_file.write(
-            """
-            import numpy as np
-
-            from astropy import units
-
-            # BEAST imports
-            from beast.physicsmodel.stars import isochrone
-            from beast.physicsmodel.stars import stellib
-            from beast.physicsmodel.dust import extinction
-            from beast.observationmodel.noisemodel import absflux_covmat
-
-            # from extra_filters import make_integration_filter, make_top_hat_filter
-
-            # -----------------------------------------------------------------
-            # User inputs                                   [sec:conf]
-            # -----------------------------------------------------------------
-            # Parameters that are required to make models
-            # and to fit the data
-            # -----------------------------------------------------------------
-            # AC == authomatically created
-            # indicates where user's input change is NOT necessary/recommeded
-            # -----------------------------------------------------------------
-
-            # project : string
-            #   the name of the output results directory
-            project = "beast_example_phat"
-
-            # name of the survey
-            #  used for the creation of the unique name for each source
-            surveyname = "PHAT"
-
-            # filters : list of strings
-            #   full filter names in BEAST filter database
-            filters = [
-                "HST_WFC3_F275W",
-                "HST_WFC3_F336W",
-                "HST_ACS_WFC_F475W",
-                "HST_ACS_WFC_F814W",
-                "HST_WFC3_F110W",
-                "HST_WFC3_F160W",
-            ]
-
-            # basefilters : list of strings
-            #   short names for filters
-            basefilters = ["F275W", "F336W", "F475W", "F814W", "F110W", "F160W"]
-
-            # obs_colnames : list of strings
-            #   names of columns for filters in the observed catalog
-            #   need to match column names in the observed catalog,
-            #   input data MUST be in fluxes, NOT in magnitudes
-            #   fluxes MUST be in normalized Vega units
-            obs_colnames = [f.lower() + "_rate" for f in basefilters]
-            # obs_colnames = [ f.upper() + '_RATE' for f in basefilters ]
-
-            # obsfile : string
-            #   pathname of the observed catalog
-            obsfile = "data/b15_4band_det_27_A.fits"
-
-            # ------------------------------------------------------
-            # Artificial Star Test Input File Generation Parameters
-            # ------------------------------------------------------
-
-            # ast_models_selected_per_age : integer
-            # Number of models to pick per age (Default = 70).
-            ast_models_selected_per_age = 70
-
-            # ast_bands_above_maglimit : integer
-            # Number of filters that must be above the magnitude limit
-            # for an AST to be included in the list (Default = 3)
-            ast_bands_above_maglimit = 3
-
-
-            # ast_realization_per_model : integer
-            # Number of Realizations of each included AST model
-            # to be put into the list. (Default = 20)
-            ast_realization_per_model = 20
-
-
-            # ast_maglimit : float (single value or array with one value per filter)
-            # (1) option 1: [number] to change the number of mags fainter than
-            #                  the 90th percentile
-            #               faintest star in the photometry catalog to be used for
-            #                  the mag cut.
-            #               (Default = 1)
-            # (2) option 2: [space-separated list of numbers] to set custom faint end limits
-            #               (one value for each band).
-            ast_maglimit = [1.0]
-
-            # ast_with_positions :  (bool,optional)
-            # If True, the ast list is produced with X,Y positions.
-            # If False, the ast list is produced with only magnitudes.
-            ast_with_positions = True
-
-            # ast_density_table :  (string,optional)
-            # Name of density table created by
-            # tools/create_background_density_map.py, containing either the source
-            # density map or the background density map. If supplied, the ASTs will
-            # be repeated for each density bin in the table
-            ast_density_table = None
-            # ast_density_table = 'data/b15_4band_det_27_A_sourcedens_map.hd5'
-
-            # ast_N_bins : (int, optional)
-            # Number of source or background bins that you want ASTs repeated over
-            # ast_N_bins = 8
-
-            # ast_pixel_distribution : float (optional)
-            # (Used if ast_with_positions is True), minimum pixel separation between AST
-            # position and catalog star used to determine the AST spatial distribution
-            ast_pixel_distribution = 10.0
-
-            # ast_reference_image : string (optional, but required if ast_with_positions
-            # is True and no X and Y information  is present in the photometry catalog)
-            # Name of the reference image used by DOLPHOT when running the measured
-            # photometry.
-            ast_reference_image = None
-
-            # ast_coord_boundary : None, or list of two arrays (optional)
-            # If supplied, these RA/Dec coordinates will be used to limit the region
-            # over which ASTs are generated.  Input should be list of two arrays, the
-            # first RA and the second Dec, ordered sequentially around the region
-            # (either CW or CCW).
-            ast_coord_boundary = None
-
-            # -------------------------------------------
-            # Noise Model Artificial Star Test Parameters
-            # -------------------------------------------
-
-            # astfile : string
-            #   pathname of the AST files (single camera ASTs)
-            astfile = "data/fake_stars_b15_27_all.hd5"
-
-            # ast_colnames : list of strings
-            #   names of columns for filters in the AST catalog (AC)
-            ast_colnames = np.array(basefilters)
-
-            # noisefile : string
-            #   create a name for the noise model
-            noisefile = project + "/" + project + "_noisemodel.grid.hd5"
-
-            # absflux calibration covariance matrix for HST specific filters (AC)
-            absflux_a_matrix = absflux_covmat.hst_frac_matrix(filters)
-
-            # -------------------------------------------
-            # Grid
-            # -------------------------------------------
-
-            # n_subgrid : integer
-            #     Number of sub-grids to use (1 means no subgrids).  These are
-            #     useful when the physics model grid is too large to read into
-            #     memory.
-            n_subgrid = 1
-
-            ################
-
-            # Distance/Velocity
-
-            # velocity of galaxy
-            velocity = -300 * units.km / units.s  # M31 velocity from SIMBAD
-
-            # Distances: distance to the galaxy [min, max, step] or [fixed number]
-            distances = [24.47]
-            # Distance unit (any length or units.mag)
-            distance_unit = units.mag
-            distance_prior_model = {'name': 'flat'}
-
-            ################
-
-            # Stellar grid definition
-
-            # log10(Age) -- [min,max,step] to generate the isochrones in years
-            #   example [6.0, 10.13, 1.0]
-            logt = [6.0, 10.13, 1.0]
-            age_prior_model = {'name': 'flat'}
-
-            # note: Mass is not sampled, instead the isochrone supplied
-            #       mass spacing is used instead
-            mass_prior_model = {"name": "kroupa"}
-
-            # Metallicity : list of floats
-            #   Here: Z == Z_initial, NOT Z(t) surface abundance
-            #   PARSECv1.2S accepts values 1.e-4 < Z < 0.06
-            #   example z = [0.03, 0.019, 0.008, 0.004]
-            #   can they be set as [min, max, step]?
-            z = [0.03, 0.019, 0.008, 0.004]
-            met_prior_model = {"name": "flat"}
-
-            #   Current Choices: Padova or MIST
-            #   PadovaWeb() -- `modeltype` param for iso sets from ezpadova
-            #      (choices: parsec12s_r14, parsec12s, 2010, 2008, 2002)
-            #   MISTWeb() -- `rotation` param (choices: vvcrit0.0=default, vvcrit0.4)
-            #
-            # Default: PARSEC+COLIBRI
-            oiso = isochrone.PadovaWeb()
-            # Alternative: PARSEC1.2S -- old grid parameters
-            # oiso = isochrone.PadovaWeb(modeltype='parsec12s', filterPMS=True)
-            # Alternative: MIST -- v1, no rotation
-            # oiso = isochrone.MISTWeb()
-
-            # Stellar Atmospheres library definition
-            osl = stellib.Tlusty() + stellib.Kurucz()
-
-            ################
-
-            # Dust extinction grid definition
-            extLaw = extinction.Gordon16_RvFALaw()
-
-            # A(V): dust column in magnitudes
-            #   acceptable avs > 0.0
-            #   example [min, max, step] = [0.0, 10.055, 1.0]
-            avs = [0.0, 10.055, 1.0]
-            av_prior_model = {"name": "flat"}
-
-            # R(V): dust average grain size
-            #   example [min, max, step] = [2.0,6.0,1.0]
-            rvs = [2.0, 6.0, 1.0]
-            rv_prior_model = {"name": "flat"}
-
-            # fA: mixture factor between "MW" and "SMCBar" extinction curves
-            #   example [min, max, step] = [0.0,1.0, 0.25]
-            fAs = [0.0, 1.0, 0.25]
-            fA_prior_model = {"name": "flat"}
-
-            ################
-
-            # add in the standard filters to enable output of stats and pdf1d values
-            # for the observed fitlers (AC)
-            add_spectral_properties_kwargs = dict(filternames=filters)
-            """
-        )
-
-    # create the settings
-    settings = beast_settings.beast_settings(temp_file.name)
-
-    # assert it's the correct class
-    assert isinstance(
-        settings, beast_settings.beast_settings
-    ), "Did not produce the correct class"
